Move the printer adapter into the ApplyTask
