+++ pkg/apply/task/printer_adapter_test.go
@@ -1,7 +1,7 @@
 // Copyright 2019 The Kubernetes Authors.
 // SPDX-License-Identifier: Apache-2.0
 
-package apply
+package task
 
 import (
 	"bytes"
